requirements.txt
discord.py==2.3.2
git+https://github.com/Rapptz/discord-ext-menus
gspread
beautifulsoup4
apscheduler